refactor : move automation related logic to a separate package
